# perf: use shared PGBoolean instances