--- a/org/postgresql/core/types/PGInteger.java
+++ b/org/postgresql/core/types/PGInteger.java
@@ -21,13 +21,22 @@ import org.postgresql.util.PSQLState;
  */
 public class PGInteger implements PGType
 {
+    static final PGInteger ZERO = new PGInteger(0);
+
+    static final PGInteger ONE = new PGInteger(1);
+
     private Integer val;
-    
+
     protected PGInteger( Integer x )
     {
         val = x;
     }
-    
+
+    static final PGInteger valueOf(final Integer value) {
+        final int v = value; // Prevent double auto-unboxing
+        return v == 0 ? PGInteger.ZERO : v == 1 ? PGInteger.ONE : new PGInteger(value);
+    }
+
     public static PGType castToServerType( Integer val, int targetType ) throws PSQLException
     {
         try
@@ -35,25 +44,25 @@ public class PGInteger implements PGType
             switch ( targetType )
             {
 	            case Types.BIT:
-	                return new PGBoolean(val == 0?Boolean.FALSE:Boolean.TRUE );
+                    return PGBoolean.valueOf(val != 0);
 	            case Types.REAL:
-	                return new PGFloat(val.floatValue());
+	                return PGFloat.valueOf(val.floatValue());
 	            case Types.DOUBLE:
 	            case Types.FLOAT:
-	                return new PGDouble(val.doubleValue());
+	                return PGDouble.valueOf(val.doubleValue());
 	            case Types.VARCHAR:
-	            case Types.LONGVARCHAR:                
-	                return new PGString( val.toString() );
+	            case Types.LONGVARCHAR:
+	                return PGString.valueOf( val.toString() );
 	            case Types.SMALLINT:
 	            case Types.TINYINT:
-	                return new PGShort(val.shortValue());
+	                return PGShort.valueOf(val.shortValue());
 	            case Types.INTEGER:
-	                return new PGInteger( val );
+	                return PGInteger.valueOf( val );
 	            case Types.DECIMAL:
 	            case Types.NUMERIC:
-	                return new PGBigDecimal( new BigDecimal( val.toString()));
+	                return PGBigDecimal.valueOf( new BigDecimal( val.toString()));
 	            default:
-	                return new PGUnknown(val);
+	                return PGUnknown.valueOf(val);
             }
         }
         catch( Exception ex )
@@ -61,6 +70,7 @@ public class PGInteger implements PGType
             throw new PSQLException(GT.tr("Cannot convert an instance of {0} to type {1}", new Object[]{val.getClass().getName(),"Types.OTHER"}), PSQLState.INVALID_PARAMETER_TYPE, ex);
         }
     }
+    @Override
     public String toString()
     {
         return val.toString();