--- a/org/postgresql/core/types/PGLong.java
+++ b/org/postgresql/core/types/PGLong.java
@@ -20,41 +20,50 @@ import org.postgresql.util.PSQLState;
  */
 public class PGLong implements PGType
 {
+    static final PGLong ZERO = new PGLong(0L);
+
+    static final PGLong ONE = new PGLong(1L);
+
     private Long val;
-    
-    protected PGLong( Long x )
+
+    private PGLong( Long x )
     {
         val = x;
     }
-    
-    public static PGType castToServerType(Long val, int targetType ) throws PSQLException 
+
+    static final PGLong valueOf(final Long value) {
+        final long v = value; // Prevent double auto-unboxing
+        return v == 0L ? PGLong.ZERO : v == 1L ? PGLong.ONE : new PGLong(value);
+    }
+
+    public static PGType castToServerType(Long val, int targetType ) throws PSQLException
     {
         try
         {
             switch ( targetType )
             {
 	            case Types.BIT:
-	                return new PGBoolean(val ==0?Boolean.FALSE:Boolean.TRUE);
+	                return PGBoolean.valueOf(val != 0L);
 	            case Types.REAL:
-	                return new PGFloat(val.floatValue());
+	                return PGFloat.valueOf(val.floatValue());
 	            case Types.FLOAT:
 	            case Types.DOUBLE:
-	                return new PGDouble(val.doubleValue());
+	                return PGDouble.valueOf(val.doubleValue());
 	            case Types.VARCHAR:
-	            case Types.LONGVARCHAR:                
-	                return new PGString(val.toString());
+	            case Types.LONGVARCHAR:
+	                return PGString.valueOf(val.toString());
 	            case Types.BIGINT:
-	                return new PGLong( val );
+	                return PGLong.valueOf(val);
 	            case Types.INTEGER:
-	                return new PGInteger(val.intValue());
+	                return PGInteger.valueOf(val.intValue());
 	            case Types.SMALLINT:
 	            case Types.TINYINT:
-	                return new PGShort(val.shortValue());
+	                return PGShort.valueOf(val.shortValue());
 	            case Types.DECIMAL:
 	            case Types.NUMERIC:
-	                return new PGBigDecimal( new BigDecimal( val.toString())); 
+	                return PGBigDecimal.valueOf( new BigDecimal( val.toString()));
 	            default:
-	                return new PGUnknown(val);
+	                return PGUnknown.valueOf(val);
             }
         }
         catch( Exception ex )
@@ -62,6 +71,7 @@ public class PGLong implements PGType
             throw new PSQLException(GT.tr("Cannot convert an instance of {0} to type {1}", new Object[]{val.getClass().getName(),"Types.OTHER"}), PSQLState.INVALID_PARAMETER_TYPE, ex);
         }
     }
+    @Override
     public String toString()
     {
         return val.toString();