--- a/org/postgresql/core/types/PGShort.java
+++ b/org/postgresql/core/types/PGShort.java
@@ -21,11 +21,22 @@ import org.postgresql.util.PSQLState;
  */
 public class PGShort implements PGType
 {
-    Short val;
-    protected  PGShort( Short x )
+    static final PGShort ZERO = new PGShort((short) 0);
+
+    static final PGShort ONE = new PGShort((short) 1);
+    
+    private Short val;
+
+    private PGShort( Short x )
     {
         val = x;
     }
+    
+    static final PGShort valueOf(final Short value) {
+        final short v = value; // Prevent double auto-unboxing
+        return v == (short) 0 ? PGShort.ZERO : v == (short) 1 ? PGShort.ONE : new PGShort(value);
+    }
+    
     public static PGType castToServerType( Short val, int targetType ) throws PSQLException
     {
         try
@@ -33,24 +44,24 @@ public class PGShort implements PGType
             switch ( targetType )
             {
 	            case Types.BIT:
-	                return new PGBoolean(val == 0?Boolean.FALSE:Boolean.TRUE );
+	                return PGBoolean.valueOf(val != (short) 0);
 	            
 	            case Types.SMALLINT:
 	            case Types.TINYINT:
-	                return new PGShort(val);
+	                return PGShort.valueOf(val);
 	            case Types.REAL:
-	                return new PGFloat(val.floatValue());
+	                return PGFloat.valueOf(val.floatValue());
 	            case Types.DOUBLE:
 	            case Types.FLOAT:
-	                return new PGDouble(val.doubleValue());
+	                return PGDouble.valueOf(val.doubleValue());
 	            case Types.VARCHAR:
 	            case Types.LONGVARCHAR:                
-	                return new PGString ( val.toString() );
+	                return PGString.valueOf( val.toString() );
 	            case Types.DECIMAL:
 	            case Types.NUMERIC:
-	                return new PGBigDecimal( new BigDecimal( val.toString()));
+	                return PGBigDecimal.valueOf( new BigDecimal( val.toString()));
 	            default:
-	                return new PGUnknown(val);
+	                return PGUnknown.valueOf(val);
 	            }
         }
         catch (Exception ex)