--- a/org/postgresql/core/types/PGType.java
+++ b/org/postgresql/core/types/PGType.java
@@ -14,5 +14,6 @@ package org.postgresql.core.types;
  */
 public interface PGType
 {
+    @Override
     public String toString();
 }