Reduce GC stress by using solely two shared PGBoolean instances.
Test shows that setObject(i, v, type) runs 13% faster for boolean.
Prepare future caching of PG* types using valueOf().

Closes #315