@mkarg 2015-06-13
Please note that I did neither check for nor fixed any *existing* potential failure, but solely implemented what I described: Prevent creation of multiple booleans. NULL is not a boolean, and looking at the code in fact I doubt that the driver is able to process is even without my PR already.