@sehrope 2015-06-13
Yeah if there's an issue it has nothing to do with this PR as it was already like that ... I'm just making a note of it to check out later as it looks odd.