@mkarg 2015-06-13
In fact I asked myself the same questions that popped up today:
* Are PG* classes useful at all?
* Will they ever have to process null?
We should certainly answer them soon -- but I think we should do things step by step.
So I filed this PR to get my idea done first, and after that, we should step on and discuss that issues in separate PRs. :-)