@vlsi 2015-06-13
>Test shows that setObject(i, v, type) runs 13% faster for boolean.

@mkarg , Can you please share the test you use to measure the performance?