@mkarg 2015-06-13
@vlsi already dropped the test source (was just a short hack). It was just a simple comparison of the execution time needed to invoke PreparedStatement.setObject(1, 1, Boolean) 100 million times (the statement was INSERT INTO Table (ID) VALUE (?)). The test runs approximately 3s with original source, and approximately 2.5s with my changes, Test was repeated 10 times.