@vlsi 2015-06-13
Please use JMH for that kind of testing and Java Flight Recorder to validate what is actually allocated.

Here the change is straightforward, but "measuring something in a loop of 100 million times" is not that good approach.