@vlsi 2015-06-13
> already dropped the test source (was just a short hack).

@mkarg, FYI. Here's how we take measurements: https://github.com/pgjdbc/pgjdbc/blob/master/ubenchmark/src/main/java/org/postgresql/benchmark/statement/ParseStatement.java

If it is worth running, it is worth placing in ubenchmark module.