@mkarg 2015-06-13
@vlsi Please Google people before. You're generally right, and sorry for that, but see, after 30+ years of programming, 17+ of them with Java daily, being head of development at an ISV for 10 years, and serving as an EG member for years (JSR 339, 370) there simply just are SOME things that I actually dare to do blindly with no further proof. Spoken frankly, this project can choose to simply accept my professional expertise as-is, or can insist on rules; don't get me wrong, but the latter will just make me volunteer somewhere else. ;-)