@vlsi 2015-06-13
> 17+ of them with Java daily

It would be nice to hear 21+ with java :)

Don't get me wrong. If I were you, I would go and change that without taking careful measurements.
As for me, current change does not even need to claim XX% gain in whatever.
However, it strikes me when I hear "100 million times".

Note: I am just a couple of weeks into pgjdbc, so I am not in a position to approve/reject contributions. I do not setup rules either.

How would you know JVM would allocate that `new PGBoolean()`? After all, JVMs do escape analysis, etc, so it might be the case the object would be stack-allocated.

Anyway, let's measure a bit.
It turns out, `ps.setObject(1, 1, Types.BOOLEAN);` and `ps.setObject(1, true, Types.BOOLEAN);` both allocate 16 bytes in 1.8u40 osx.

Your fix indeed improves that down to zero (no surprise here).
The interesting part is there is *no visible gain in response time* when running single-threaded benchmark.

I do not mean the change is bad. I just draw a conclusion out of a valid measurement.

```java
ps = connection.prepareStatement("select ?");

@Benchmark
public Statement boolAsInt() throws SQLException {
    ps.setObject(1, 1, Types.BOOLEAN);
    return ps;
}

@Benchmark
public Statement boolAsBoolean() throws SQLException {
    ps.setObject(1, true, Types.BOOLEAN);
    return ps;
}

@Benchmark
public Statement bindBoolean() throws SQLException {
    ps.setBoolean(1, true);
    return ps;
}
```


### Before

```
Benchmark                                      Mode  Cnt    Score    Error   Units
BindBoolean.bindBoolean                        avgt   10    7,792 ±  1,104   ns/op
BindBoolean.bindBoolean:·gc.alloc.rate.norm    avgt   10   ≈ 10⁻⁵             B/op
BindBoolean.boolAsBoolean                      avgt   10   23,594 ±  1,541   ns/op
BindBoolean.boolAsBoolean:·gc.alloc.rate.norm  avgt   10   16,000 ±  0,001    B/op
BindBoolean.boolAsInt                          avgt   10   15,580 ±  0,863   ns/op
BindBoolean.boolAsInt:·gc.alloc.rate.norm      avgt   10   16,000 ±  0,001    B/op
```

### After

```
Benchmark                                      Mode  Cnt   Score    Error   Units
BindBoolean.bindBoolean                        avgt   10   7,663 ±  0,353   ns/op
BindBoolean.boolAsBoolean                      avgt   10  21,849 ±  2,502   ns/op
BindBoolean.boolAsInt                          avgt   10  15,241 ±  1,839   ns/op
```


Full benchmark code for the reference
```java
package org.postgresql.benchmark.statement;

import org.openjdk.jmh.annotations.*;
import org.openjdk.jmh.profile.GCProfiler;
import org.openjdk.jmh.runner.Runner;
import org.openjdk.jmh.runner.RunnerException;
import org.openjdk.jmh.runner.options.Options;
import org.openjdk.jmh.runner.options.OptionsBuilder;
import org.postgresql.benchmark.profilers.FlightRecorderProfiler;
import org.postgresql.util.ConnectionUtil;

import java.sql.*;
import java.util.Properties;
import java.util.concurrent.TimeUnit;

@Fork(value = 1, jvmArgsPrepend = "-Xmx128m")
@Measurement(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
@Warmup(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
@State(Scope.Thread)
@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
public class BindBoolean {
    private Connection connection;
    private PreparedStatement ps;

    @Setup(Level.Trial)
    public void setUp() throws SQLException {
        Properties props = ConnectionUtil.getProperties();

        connection = DriverManager.getConnection(ConnectionUtil.getURL(), props);
        ps = connection.prepareStatement("select ?");
    }

    @TearDown(Level.Trial)
    public void tearDown() throws SQLException {
        ps.close();
        connection.close();
    }

    @Benchmark
    public Statement boolAsInt() throws SQLException {
        ps.setObject(1, 1, Types.BOOLEAN);
        return ps;
    }

    @Benchmark
    public Statement boolAsBoolean() throws SQLException {
        ps.setObject(1, true, Types.BOOLEAN);
        return ps;
    }

    @Benchmark
    public Statement bindBoolean() throws SQLException {
        ps.setBoolean(1, true);
        return ps;
    }

    public static void main(String[] args) throws RunnerException {
        Options opt = new OptionsBuilder()
                .include(BindBoolean.class.getSimpleName())
                .addProfiler(GCProfiler.class)
                .detectJvmArgs()
                .build();

        new Runner(opt).run();
    }
}
```